Pull out coordinate code to separate file
